Remove unused functions for handling email-sign-in
